+++ General/VideoScript.vb
@@ -236,7 +236,6 @@ clipname.set_output()
                     ModifyScript(code, Engine).WriteFile(Path, Encoding.UTF8)
                 Else
                     ModifyScript(code, Engine).WriteFileDefault(Path)
-                    Environment.SetEnvironmentVariable("AviSynthDLL", Package.AviSynth.Path)
                 End If
 
                 If Not Package.AviSynth.VerifyOK OrElse